1、优化代码：建立命令类 2、优化代码：建立工具类
