+++ RemoteCtr/RemoteCtr/MyTool.h
@@ -1,0 +1,17 @@
+#pragma once
+class CMyTool
+{
+public:
+	static void Dump(BYTE* pData, size_t nSize) {
+        std::string strout;
+        for (size_t i = 0; i < nSize; i++) {
+            char buf[8] = "";
+            if (i > 0 && (i % 16 == 0))strout += "\n";
+            snprintf(buf, sizeof(buf), "%02X ", pData[i] & 0xFF);
+            strout += buf;
+        }
+        strout += "\n";
+        OutputDebugStringA(strout.c_str());
+    }
+};
+
