+++ RemoteCtr/RemoteCtr/RemoteCtr.vcxproj.filters
@@ -36,6 +36,12 @@
     <ClInclude Include="LockDialog.h">
       <Filter>头文件</Filter>
     </ClInclude>
+    <ClInclude Include="Command.h">
+      <Filter>头文件</Filter>
+    </ClInclude>
+    <ClInclude Include="MyTool.h">
+      <Filter>头文件</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="RemoteCtr.cpp">
@@ -48,6 +54,12 @@
       <Filter>源文件</Filter>
     </ClCompile>
     <ClCompile Include="LockDialog.cpp">
+      <Filter>源文件</Filter>
+    </ClCompile>
+    <ClCompile Include="Command.cpp">
+      <Filter>源文件</Filter>
+    </ClCompile>
+    <ClCompile Include="MyTool.cpp">
       <Filter>源文件</Filter>
     </ClCompile>
   </ItemGroup>
